Migrated the two API calls containing slugs out of the Models Added the function that filters the show/archive episodes It doesn't send episodes which gonna be premiered
